delete endpoint

+++ webapp/cloud-ninja/src/main/java/edu/neu/csye6225/spring19/cloudninja/service/impl/LoginServiceImpl.java
@@ -1,5 +1,8 @@
 package edu.neu.csye6225.spring19.cloudninja.service.impl;
 
+import static edu.neu.csye6225.spring19.cloudninja.constants.ApplicationConstants.ADMIN_EMAIL;
+import static edu.neu.csye6225.spring19.cloudninja.constants.ApplicationConstants.ADMIN_PASSWD;
+
 import java.text.SimpleDateFormat;
 import java.util.List;
 
@@ -8,6 +11,7 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import edu.neu.csye6225.spring19.cloudninja.exception.ResourceNotFoundException;
 import edu.neu.csye6225.spring19.cloudninja.exception.UnAuthorizedLoginException;
 import edu.neu.csye6225.spring19.cloudninja.exception.ValidationException;
 import edu.neu.csye6225.spring19.cloudninja.model.ResponseBody;
@@ -83,10 +87,21 @@ public class LoginServiceImpl implements LoginService {
 	}
 
 	@Override
-	public void deleteAllUsers(String auth) throws ValidationException {
+	public void deleteAllUsers(String auth)
+			throws ValidationException, UnAuthorizedLoginException, ResourceNotFoundException {
 
 		try {
-			userRepository.deleteAll();
+			UserCredentials credentials = authService.extractCredentialsFromHeader(auth);
+			authService.authenticateUser(credentials);
+			if (credentials.getEmailId().equals(ADMIN_EMAIL)) {
+				userRepository.deleteAll();
+				UserCredentials creds = new UserCredentials();
+				creds.setEmailId(ADMIN_EMAIL);
+				creds.setPassword(authServiceUtil.encryptPassword(ADMIN_PASSWD));
+				userRepository.save(creds);
+			} else {
+				throw new UnAuthorizedLoginException("Not an admin role. Can't delete the database.");
+			}
 		} catch (Exception e) {
 			logger.error(commonUtil.stackTraceString(e));
 			throw e;
